Feb 20, 2026, 9:59 AM

+++ DNAStrand.cpp
@@ -52,7 +52,7 @@ DNAStrand::DNAStrand(int lengthValue)
 }
 
 DNAStrand::DNAStrand(const std::string &startingString) {
-  m_length = startingString.size();
+  m_length = static_cast<int>(startingString.size());
   m_bases  = new Base[m_length];
   for (int i = 0; i < m_length; ++i) {
     m_bases[i] = charToBase(startingString.at(i));
